+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "noteour" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.1]
 
 - Initial release
+
+## [0.0.2]
+
+- Fixes: 
+  - Updated core functionality to support special charecters in note type, such as '@', etc.
+- Updated the naming of CMD+SHIFT+P(CTRL+SHIFT+P) commands to be more descriptive.
+- Updated the core functionality to-collect audit notes upon each file extension or note-type change.